implemented tests for delete by id not found for beer and customer

+++ src/main/java/marc/springframework/spring6restmvc/services/BeerServiceJPA.java
@@ -62,9 +62,14 @@ public class BeerServiceJPA implements BeerService
     }
 
     @Override
-    public void deleteById(UUID beerId)
+    public Boolean deleteById(UUID beerId)
     {
-        beerRepository.deleteById(beerId);
+        if (beerRepository.existsById(beerId))
+        {
+            beerRepository.deleteById(beerId);
+            return true;
+        }
+        return false;
     }
 
     @Override
