+++ src/main/java/marc/springframework/spring6restmvc/services/CustomerServiceJPA.java
@@ -59,9 +59,14 @@ public class CustomerServiceJPA implements CustomerService
     }
 
     @Override
-    public void deleteById(UUID customerId)
+    public Boolean deleteById(UUID customerId)
     {
-        customerRepository.deleteById(customerId);
+        if (customerRepository.existsById(customerId))
+        {
+            customerRepository.deleteById(customerId);
+            return true;
+        }
+        return false;
     }
 
     @Override
